Fold jq explosion tests into transform_test.go
The package keeps all transformer tests in one transform_test.go; the
separate jq_multi_test.go was only split out to dodge in-flight parallel
edits. Move both tests next to the existing TestJq and drop the file.
